Create README.md
This project is for my Final in  SDEV140, I have chosen to create an app/program for ordering food(specifically) sandwiches, chips and drinks.  The user will have options in each of these three categories, and will be able to input a quantity they want to order for each item.  The program will confirm the items ordered, and display their order total for confirmation before placing the order.